sidq-backend/requirements.txt
alembic==1.15.2
astroid==3.3.8
backports.tarfile==1.2.0
blinker==1.9.0
branca==0.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cpi==2.0.4
dill==0.3.9
docutils==0.21.2
flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
folium==0.19.5
greenlet==3.1.1
hererocks==0.25.1
idna==3.10
importlib-metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jinja2==3.1.5
keyring==25.5.0
mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.19
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pkginfo==1.12.0
platformdirs==4.3.6
psycopg2==2.9.10
pygments==2.18.0
PyJWT==2.10.1
pylint==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
readme-renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
six==1.16.0
SQLAlchemy==2.0.37
tomli==2.2.1
tomlkit==0.13.2
twine==6.0.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.1.3
xyzservices==2025.1.0
zipp==3.21.0
